pyproject.toml
[project]
name = "llm-orchestrater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.30.1",
    "jinja2>=3.1.6",
    "litellm>=1.76.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "stripe>=12.4.0",
    "temporalio>=1.16.0",
    "uvicorn>=0.35.0",
]
